recommendation-service/requirements.txt
fastapi
uvicorn
tensorflow-cpu
numpy
pandas
scikit-learn
pydantic